Fix delete confirmation
Shouldn't have done this in this branch.

+++ resources/assets/coffee/react/_components/comment.coffee
@@ -137,7 +137,6 @@ class @Comment extends React.PureComponent
                   type: 'button'
                   className: 'comment__action'
                   onClick: @delete
-                  'data-confirm': osu.trans('common.confirmation')
                   osu.trans('common.buttons.delete')
 
             if @props.comment.replies_count > 0
@@ -212,6 +211,8 @@ class @Comment extends React.PureComponent
 
 
   delete: =>
+    return unless confirm(osu.trans('common.confirmation'))
+
     $.ajax laroute.route('comments.destroy', comment: @props.comment.id),
       method: 'DELETE'
     .done (data) =>
